+++ manage.py
@@ -8,6 +8,7 @@ import click
 from src.command.find_samples import find_samples
 from src.command.upload_photos import upload_photos
 from src.command.process_photos import process_photos
+from src.command.build import build
 
 
 @click.group()
@@ -20,6 +21,7 @@ def cli():
 cli.add_command(find_samples, name="find-samples")
 cli.add_command(upload_photos, name="upload-photos")
 cli.add_command(process_photos, name="process-photos")
+cli.add_command(build, name="build")
 
 
 if __name__ == "__main__":